+++ packages/zoho/src/Adaptor.js
@@ -1,7 +1,5 @@
-import {
-  execute as commonExecute,
-  expandReferences,
-} from '@openfn/language-common';
+import { execute as commonExecute } from '@openfn/language-common';
+import { expandReferences } from '@openfn/language-common/util';
 import { post } from './Client';
 
 /**
@@ -44,7 +42,7 @@ export function execute(...operations) {
 export function addRow(db, table, rowData) {
   return state => {
     const action = 'ADDROW';
-    const body = expandReferences(rowData)(state);
+    const [body] = expandReferences(state, rowData);
 
     const { account, authToken, apiVersion } = state.configuration;