+++ packages/fhir-ndr-et/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @openfn/language-fhir-ndr-et
 
+## 0.1.16 - 10 July 2025
+
+### Patch Changes
+
+- Updated dependencies \[ea85695]
+- Updated dependencies \[3fce58f]
+- Updated dependencies \[19f2d7e]
+- Updated dependencies \[f26bd2b]
+- Updated dependencies \[19f2d7e]
+  - @openfn/language-common@3.0.0
+  - @openfn/language-fhir@5.0.5
+
 ## 0.1.15 - 20 June 2025
 
 ### Patch Changes